Export the vm interface
Add documentation for the interfaces
