+++ internal/vm/libkrun/krun.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 	"unsafe"
 
-	"github.com/containerd/nerdbox/internal/vm"
+	"github.com/containerd/nerdbox/pkg/vm"
 )
 
 const (
