+++ plugins/shim/sandbox/plugin.go
@@ -21,7 +21,7 @@ import (
 	"github.com/containerd/plugin/registry"
 
 	vmsbox "github.com/containerd/nerdbox/internal/shim/sandbox/vm"
-	"github.com/containerd/nerdbox/internal/vm"
+	"github.com/containerd/nerdbox/pkg/vm"
 	"github.com/containerd/nerdbox/plugins"
 )
